+++ src/admin/pages/map-styles.tsx
@@ -1,7 +1,5 @@
-import { useForm } from "@tanstack/react-form";
 import {
   BaseControl,
-  Button,
   Icon,
   PanelBody,
   TextControl,
@@ -11,7 +9,7 @@ import * as z from "zod/mini";
 
 import { zUptrackSettings } from "../../settings";
 import { useUpdateSettings } from "../use-update-settings";
-import { FormSubmitNotice, mountAdminPage } from "./shared";
+import { FormSubmitNotice, mountAdminPage, useAdminForm } from "./_shared";
 
 const parseNumberInput = (value: string): number =>
   Number.parseFloat(value) || 0;
@@ -54,7 +52,7 @@ function MapStylesPage({
 }): React.JSX.Element {
   const { result, update } = useUpdateSettings();
 
-  const form = useForm({
+  const form = useAdminForm({
     defaultValues: settingsDefault.uptrack_map_styles,
     onSubmit: async ({ value }) => {
       await update({ uptrack_map_styles: value });
@@ -62,270 +60,29 @@ function MapStylesPage({
   });
 
   return (
-    <form
-      className="form-wrap"
-      onSubmit={(event) => {
-        event.preventDefault();
-        event.stopPropagation();
-        void form.handleSubmit();
-      }}
-    >
-      <FormSubmitNotice result={result} />
+    <form.AppForm>
+      <form
+        className="form-wrap"
+        onSubmit={(event) => {
+          event.preventDefault();
+          event.stopPropagation();
+          void form.handleSubmit();
+        }}
+      >
+        <FormSubmitNotice result={result} />
 
-      <div className="form-field">
-        <BaseControl label="Map Styles">
-          <PanelBody title="Map Styles" initialOpen={false}>
-            <div className="map-styles-field-group">
-              <form.Field
-                name="canvasTolerance"
-                children={(field) => (
-                  <TextControl
-                    label={withHelpLabel(
-                      "Canvas Tolerance (px)",
-                      "Pixel tolerance for canvas hit detection. Higher values make line hover/click interactions more forgiving.",
-                    )}
-                    type="number"
-                    min={0}
-                    step="any"
-                    value={field.state.value.toString()}
-                    onChange={(value) => {
-                      field.handleChange(parseNumberInput(value));
-                    }}
-                    __next40pxDefaultSize
-                    __nextHasNoMarginBottom
-                  />
-                )}
-              />
-              <form.Field
-                name="roundtripEpsilon"
-                children={(field) => (
-                  <TextControl
-                    label={withHelpLabel(
-                      "Roundtrip Epsilon",
-                      "Distance threshold used to treat route start and end points as the same location.",
-                    )}
-                    type="number"
-                    min={0}
-                    step="any"
-                    value={field.state.value.toString()}
-                    onChange={(value) => {
-                      field.handleChange(parseNumberInput(value));
-                    }}
-                    __next40pxDefaultSize
-                    __nextHasNoMarginBottom
-                  />
-                )}
-              />
-            </div>
-
-            <div>
-              <h4>Route Styles</h4>
-              <table className="map-styles-table">
-                <thead>
-                  <tr>
-                    <th>Variant</th>
-                    <th>Opacity</th>
-                    <th>Weight</th>
-                  </tr>
-                </thead>
-                <tbody>
-                  <tr>
-                    <td>Normal</td>
-                    <td>
-                      <form.Field
-                        name="routeStyles.normal.opacity"
-                        children={(field) => (
-                          <TextControl
-                            type="number"
-                            min={0}
-                            step="any"
-                            value={field.state.value.toString()}
-                            onChange={(value) => {
-                              field.handleChange(parseNumberInput(value));
-                            }}
-                            __next40pxDefaultSize
-                            __nextHasNoMarginBottom
-                          />
-                        )}
-                      />
-                    </td>
-                    <td>
-                      <form.Field
-                        name="routeStyles.normal.weight"
-                        children={(field) => (
-                          <TextControl
-                            type="number"
-                            min={0}
-                            step="any"
-                            value={field.state.value.toString()}
-                            onChange={(value) => {
-                              field.handleChange(parseNumberInput(value));
-                            }}
-                            __next40pxDefaultSize
-                            __nextHasNoMarginBottom
-                          />
-                        )}
-                      />
-                    </td>
-                  </tr>
-                  <tr>
-                    <td>Focus</td>
-                    <td>
-                      <form.Field
-                        name="routeStyles.focus.opacity"
-                        children={(field) => (
-                          <TextControl
-                            type="number"
-                            min={0}
-                            step="any"
-                            value={field.state.value.toString()}
-                            onChange={(value) => {
-                              field.handleChange(parseNumberInput(value));
-                            }}
-                            __next40pxDefaultSize
-                            __nextHasNoMarginBottom
-                          />
-                        )}
-                      />
-                    </td>
-                    <td>
-                      <form.Field
-                        name="routeStyles.focus.weight"
-                        children={(field) => (
-                          <TextControl
-                            type="number"
-                            min={0}
-                            step="any"
-                            value={field.state.value.toString()}
-                            onChange={(value) => {
-                              field.handleChange(parseNumberInput(value));
-                            }}
-                            __next40pxDefaultSize
-                            __nextHasNoMarginBottom
-                          />
-                        )}
-                      />
-                    </td>
-                  </tr>
-                  <tr>
-                    <td>Fade</td>
-                    <td>
-                      <form.Field
-                        name="routeStyles.fade.opacity"
-                        children={(field) => (
-                          <TextControl
-                            type="number"
-                            min={0}
-                            step="any"
-                            value={field.state.value.toString()}
-                            onChange={(value) => {
-                              field.handleChange(parseNumberInput(value));
-                            }}
-                            __next40pxDefaultSize
-                            __nextHasNoMarginBottom
-                          />
-                        )}
-                      />
-                    </td>
-                    <td>
-                      <form.Field
-                        name="routeStyles.fade.weight"
-                        children={(field) => (
-                          <TextControl
-                            type="number"
-                            min={0}
-                            step="any"
-                            value={field.state.value.toString()}
-                            onChange={(value) => {
-                              field.handleChange(parseNumberInput(value));
-                            }}
-                            __next40pxDefaultSize
-                            __nextHasNoMarginBottom
-                          />
-                        )}
-                      />
-                    </td>
-                  </tr>
-                </tbody>
-              </table>
-            </div>
-
-            <div>
-              <h4>Route Type Colors</h4>
-              <table className="map-styles-table">
-                <thead>
-                  <tr>
-                    <th>Route Type</th>
-                    <th>Color</th>
-                  </tr>
-                </thead>
-                <tbody>
-                  <tr>
-                    <td>Ski Touring</td>
-                    <td>
-                      <form.Field
-                        name="routeTypeProps.ski_touring.color"
-                        children={(field) => (
-                          <TextControl
-                            value={field.state.value}
-                            onChange={(value) => {
-                              field.handleChange(value);
-                            }}
-                            __next40pxDefaultSize
-                            __nextHasNoMarginBottom
-                          />
-                        )}
-                      />
-                    </td>
-                  </tr>
-                  <tr>
-                    <td>Mountaineering</td>
-                    <td>
-                      <form.Field
-                        name="routeTypeProps.mountaineering.color"
-                        children={(field) => (
-                          <TextControl
-                            value={field.state.value}
-                            onChange={(value) => {
-                              field.handleChange(value);
-                            }}
-                            __next40pxDefaultSize
-                            __nextHasNoMarginBottom
-                          />
-                        )}
-                      />
-                    </td>
-                  </tr>
-                  <tr>
-                    <td>Hiking</td>
-                    <td>
-                      <form.Field
-                        name="routeTypeProps.hiking.color"
-                        children={(field) => (
-                          <TextControl
-                            value={field.state.value}
-                            onChange={(value) => {
-                              field.handleChange(value);
-                            }}
-                            __next40pxDefaultSize
-                            __nextHasNoMarginBottom
-                          />
-                        )}
-                      />
-                    </td>
-                  </tr>
-                </tbody>
-              </table>
-            </div>
-
-            <div>
-              <h4>Marker Styling</h4>
-              <div className="map-styles-field-group map-styles-field-group--compact map-styles-field-group--single-line">
+        <div className="form-field">
+          <BaseControl label="Map Styles">
+            <PanelBody title="Map Styles" initialOpen={false}>
+              <div className="map-styles-field-group">
                 <form.Field
-                  name="markerRadiusPx"
+                  name="canvasTolerance"
                   children={(field) => (
                     <TextControl
-                      label="Radius (px)"
+                      label={withHelpLabel(
+                        "Canvas Tolerance (px)",
+                        "Pixel tolerance for canvas hit detection. Higher values make line hover/click interactions more forgiving.",
+                      )}
                       type="number"
                       min={0}
                       step="any"
@@ -339,89 +96,19 @@ function MapStylesPage({
                   )}
                 />
                 <form.Field
-                  name="markerWeightPx"
+                  name="roundtripEpsilon"
                   children={(field) => (
                     <TextControl
-                      label="Weight (px)"
+                      label={withHelpLabel(
+                        "Roundtrip Epsilon",
+                        "Distance threshold used to treat route start and end points as the same location.",
+                      )}
                       type="number"
                       min={0}
                       step="any"
                       value={field.state.value.toString()}
                       onChange={(value) => {
                         field.handleChange(parseNumberInput(value));
-                      }}
-                      __next40pxDefaultSize
-                      __nextHasNoMarginBottom
-                    />
-                  )}
-                />
-                <form.Field
-                  name="markerColor"
-                  children={(field) => (
-                    <TextControl
-                      label="Border Color"
-                      value={field.state.value}
-                      onChange={(value) => {
-                        field.handleChange(value);
-                      }}
-                      __next40pxDefaultSize
-                      __nextHasNoMarginBottom
-                    />
-                  )}
-                />
-                <form.Field
-                  name="markerFillOpacity"
-                  children={(field) => (
-                    <TextControl
-                      label="Fill Opacity"
-                      type="number"
-                      min={0}
-                      step="any"
-                      value={field.state.value.toString()}
-                      onChange={(value) => {
-                        field.handleChange(parseNumberInput(value));
-                      }}
-                      __next40pxDefaultSize
-                      __nextHasNoMarginBottom
-                    />
-                  )}
-                />
-                <form.Field
-                  name="markerStartFillColor"
-                  children={(field) => (
-                    <TextControl
-                      label="Start Fill Color"
-                      value={field.state.value}
-                      onChange={(value) => {
-                        field.handleChange(value);
-                      }}
-                      __next40pxDefaultSize
-                      __nextHasNoMarginBottom
-                    />
-                  )}
-                />
-                <form.Field
-                  name="markerEndFillColor"
-                  children={(field) => (
-                    <TextControl
-                      label="End Fill Color"
-                      value={field.state.value}
-                      onChange={(value) => {
-                        field.handleChange(value);
-                      }}
-                      __next40pxDefaultSize
-                      __nextHasNoMarginBottom
-                    />
-                  )}
-                />
-                <form.Field
-                  name="markerRoundtripFillColor"
-                  children={(field) => (
-                    <TextControl
-                      label="Roundtrip Fill Color"
-                      value={field.state.value}
-                      onChange={(value) => {
-                        field.handleChange(value);
                       }}
                       __next40pxDefaultSize
                       __nextHasNoMarginBottom
@@ -429,25 +116,325 @@ function MapStylesPage({
                   )}
                 />
               </div>
-            </div>
-          </PanelBody>
-        </BaseControl>
-      </div>
 
-      <form.Subscribe
-        selector={(state) => [state.canSubmit, state.isSubmitting]}
-        children={([canSubmit, isSubmitting]) => (
-          <Button
-            variant="primary"
-            type="submit"
-            disabled={!canSubmit || isSubmitting}
-            isBusy={isSubmitting}
-            __next40pxDefaultSize
-          >
-            Save
-          </Button>
-        )}
-      />
-    </form>
+              <div>
+                <h4>Route Styles</h4>
+                <table className="map-styles-table">
+                  <thead>
+                    <tr>
+                      <th>Variant</th>
+                      <th>Opacity</th>
+                      <th>Weight</th>
+                    </tr>
+                  </thead>
+                  <tbody>
+                    <tr>
+                      <td>Normal</td>
+                      <td>
+                        <form.Field
+                          name="routeStyles.normal.opacity"
+                          children={(field) => (
+                            <TextControl
+                              type="number"
+                              min={0}
+                              step="any"
+                              value={field.state.value.toString()}
+                              onChange={(value) => {
+                                field.handleChange(parseNumberInput(value));
+                              }}
+                              __next40pxDefaultSize
+                              __nextHasNoMarginBottom
+                            />
+                          )}
+                        />
+                      </td>
+                      <td>
+                        <form.Field
+                          name="routeStyles.normal.weight"
+                          children={(field) => (
+                            <TextControl
+                              type="number"
+                              min={0}
+                              step="any"
+                              value={field.state.value.toString()}
+                              onChange={(value) => {
+                                field.handleChange(parseNumberInput(value));
+                              }}
+                              __next40pxDefaultSize
+                              __nextHasNoMarginBottom
+                            />
+                          )}
+                        />
+                      </td>
+                    </tr>
+                    <tr>
+                      <td>Focus</td>
+                      <td>
+                        <form.Field
+                          name="routeStyles.focus.opacity"
+                          children={(field) => (
+                            <TextControl
+                              type="number"
+                              min={0}
+                              step="any"
+                              value={field.state.value.toString()}
+                              onChange={(value) => {
+                                field.handleChange(parseNumberInput(value));
+                              }}
+                              __next40pxDefaultSize
+                              __nextHasNoMarginBottom
+                            />
+                          )}
+                        />
+                      </td>
+                      <td>
+                        <form.Field
+                          name="routeStyles.focus.weight"
+                          children={(field) => (
+                            <TextControl
+                              type="number"
+                              min={0}
+                              step="any"
+                              value={field.state.value.toString()}
+                              onChange={(value) => {
+                                field.handleChange(parseNumberInput(value));
+                              }}
+                              __next40pxDefaultSize
+                              __nextHasNoMarginBottom
+                            />
+                          )}
+                        />
+                      </td>
+                    </tr>
+                    <tr>
+                      <td>Fade</td>
+                      <td>
+                        <form.Field
+                          name="routeStyles.fade.opacity"
+                          children={(field) => (
+                            <TextControl
+                              type="number"
+                              min={0}
+                              step="any"
+                              value={field.state.value.toString()}
+                              onChange={(value) => {
+                                field.handleChange(parseNumberInput(value));
+                              }}
+                              __next40pxDefaultSize
+                              __nextHasNoMarginBottom
+                            />
+                          )}
+                        />
+                      </td>
+                      <td>
+                        <form.Field
+                          name="routeStyles.fade.weight"
+                          children={(field) => (
+                            <TextControl
+                              type="number"
+                              min={0}
+                              step="any"
+                              value={field.state.value.toString()}
+                              onChange={(value) => {
+                                field.handleChange(parseNumberInput(value));
+                              }}
+                              __next40pxDefaultSize
+                              __nextHasNoMarginBottom
+                            />
+                          )}
+                        />
+                      </td>
+                    </tr>
+                  </tbody>
+                </table>
+              </div>
+
+              <div>
+                <h4>Route Type Colors</h4>
+                <table className="map-styles-table">
+                  <thead>
+                    <tr>
+                      <th>Route Type</th>
+                      <th>Color</th>
+                    </tr>
+                  </thead>
+                  <tbody>
+                    <tr>
+                      <td>Ski Touring</td>
+                      <td>
+                        <form.Field
+                          name="routeTypeProps.ski_touring.color"
+                          children={(field) => (
+                            <TextControl
+                              value={field.state.value}
+                              onChange={(value) => {
+                                field.handleChange(value);
+                              }}
+                              __next40pxDefaultSize
+                              __nextHasNoMarginBottom
+                            />
+                          )}
+                        />
+                      </td>
+                    </tr>
+                    <tr>
+                      <td>Mountaineering</td>
+                      <td>
+                        <form.Field
+                          name="routeTypeProps.mountaineering.color"
+                          children={(field) => (
+                            <TextControl
+                              value={field.state.value}
+                              onChange={(value) => {
+                                field.handleChange(value);
+                              }}
+                              __next40pxDefaultSize
+                              __nextHasNoMarginBottom
+                            />
+                          )}
+                        />
+                      </td>
+                    </tr>
+                    <tr>
+                      <td>Hiking</td>
+                      <td>
+                        <form.Field
+                          name="routeTypeProps.hiking.color"
+                          children={(field) => (
+                            <TextControl
+                              value={field.state.value}
+                              onChange={(value) => {
+                                field.handleChange(value);
+                              }}
+                              __next40pxDefaultSize
+                              __nextHasNoMarginBottom
+                            />
+                          )}
+                        />
+                      </td>
+                    </tr>
+                  </tbody>
+                </table>
+              </div>
+
+              <div>
+                <h4>Marker Styling</h4>
+                <div className="map-styles-field-group map-styles-field-group--compact map-styles-field-group--single-line">
+                  <form.Field
+                    name="markerRadiusPx"
+                    children={(field) => (
+                      <TextControl
+                        label="Radius (px)"
+                        type="number"
+                        min={0}
+                        step="any"
+                        value={field.state.value.toString()}
+                        onChange={(value) => {
+                          field.handleChange(parseNumberInput(value));
+                        }}
+                        __next40pxDefaultSize
+                        __nextHasNoMarginBottom
+                      />
+                    )}
+                  />
+                  <form.Field
+                    name="markerWeightPx"
+                    children={(field) => (
+                      <TextControl
+                        label="Weight (px)"
+                        type="number"
+                        min={0}
+                        step="any"
+                        value={field.state.value.toString()}
+                        onChange={(value) => {
+                          field.handleChange(parseNumberInput(value));
+                        }}
+                        __next40pxDefaultSize
+                        __nextHasNoMarginBottom
+                      />
+                    )}
+                  />
+                  <form.Field
+                    name="markerColor"
+                    children={(field) => (
+                      <TextControl
+                        label="Border Color"
+                        value={field.state.value}
+                        onChange={(value) => {
+                          field.handleChange(value);
+                        }}
+                        __next40pxDefaultSize
+                        __nextHasNoMarginBottom
+                      />
+                    )}
+                  />
+                  <form.Field
+                    name="markerFillOpacity"
+                    children={(field) => (
+                      <TextControl
+                        label="Fill Opacity"
+                        type="number"
+                        min={0}
+                        step="any"
+                        value={field.state.value.toString()}
+                        onChange={(value) => {
+                          field.handleChange(parseNumberInput(value));
+                        }}
+                        __next40pxDefaultSize
+                        __nextHasNoMarginBottom
+                      />
+                    )}
+                  />
+                  <form.Field
+                    name="markerStartFillColor"
+                    children={(field) => (
+                      <TextControl
+                        label="Start Fill Color"
+                        value={field.state.value}
+                        onChange={(value) => {
+                          field.handleChange(value);
+                        }}
+                        __next40pxDefaultSize
+                        __nextHasNoMarginBottom
+                      />
+                    )}
+                  />
+                  <form.Field
+                    name="markerEndFillColor"
+                    children={(field) => (
+                      <TextControl
+                        label="End Fill Color"
+                        value={field.state.value}
+                        onChange={(value) => {
+                          field.handleChange(value);
+                        }}
+                        __next40pxDefaultSize
+                        __nextHasNoMarginBottom
+                      />
+                    )}
+                  />
+                  <form.Field
+                    name="markerRoundtripFillColor"
+                    children={(field) => (
+                      <TextControl
+                        label="Roundtrip Fill Color"
+                        value={field.state.value}
+                        onChange={(value) => {
+                          field.handleChange(value);
+                        }}
+                        __next40pxDefaultSize
+                        __nextHasNoMarginBottom
+                      />
+                    )}
+                  />
+                </div>
+              </div>
+            </PanelBody>
+          </BaseControl>
+        </div>
+
+        <form.SubmitButton />
+      </form>
+    </form.AppForm>
   );
 }
